@@ -5,13 +5,13 @@ This module provides the foundational data structures used throughout the
 dependency analysis and documentation generation system.
 """
 
-from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Optional
+
+from jinja2 import Environment, FileSystemLoader
+from pydantic import BaseModel, Field
 
 
-@dataclass
-class ImportStatement:
+class ImportStatement(BaseModel):
     """
     Represents a single import statement extracted from Python code.
 
@@ -24,15 +24,25 @@ class ImportStatement:
         is_relative: True if this is a relative import (e.g., 'from . import x')
         relative_level: Number of dots in relative import (e.g., 2 for 'from .. import x')
     """
+
     module_name: str
-    alias: Optional[str] = None
+    alias: str | None = None
     from_import: bool = False
-    imported_items: list[str] = field(default_factory=list)
+    imported_items: list[str] = Field(default_factory=list)
     line_number: int = 0
     is_relative: bool = False
     relative_level: int = 0
 
+    def __str__(self) -> str:
+        """Return a string representation of the import statement."""
+        if self.from_import:
+            items = ", ".join(self.imported_items) if self.imported_items else "*"
+            return f"from {self.module_name} import {items}"
+        alias_part = f" as {self.alias}" if self.alias else ""
+        return f"import {self.module_name}{alias_part}"
+
     def __repr__(self) -> str:
+        """Return a detailed representation for debugging."""
         if self.from_import:
             items = ", ".join(self.imported_items) if self.imported_items else "*"
             return f"from {self.module_name} import {items}"
@@ -40,8 +50,7 @@ class ImportStatement:
         return f"import {self.module_name}{alias_part}"
 
 
-@dataclass
-class ModuleNode:
+class ModuleNode(BaseModel):
     """
     Enhanced ModuleNode representing a Python module or package with dependency tracking.
 
@@ -73,49 +82,52 @@ class ModuleNode:
         cycle_group: list of ModuleNode objects in the same circular dependency group
         is_cycle_representative: True if this node represents its cycle group
     """
+
+    model_config = {"arbitrary_types_allowed": True}
+
     # Basic module information
     path: str
     name: str
     is_package: bool = False
     is_root: bool = False
-    children: list['ModuleNode'] = field(default_factory=list)
-    content: Optional[str] = None
-    documentation: Optional[str] = None
+    children: list["ModuleNode"] = Field(default_factory=list)
+    content: str | None = None
+    documentation: str | None = None
     processed: bool = False
 
     # Dependency tracking
-    dependencies: list['ModuleNode'] = field(default_factory=list)
-    dependents: list['ModuleNode'] = field(default_factory=list)
-    import_statements: list[ImportStatement] = field(default_factory=list)
-    resolved_imports: dict[str, str] = field(default_factory=dict)
+    dependencies: list["ModuleNode"] = Field(default_factory=list)
+    dependents: list["ModuleNode"] = Field(default_factory=list)
+    import_statements: list[ImportStatement] = Field(default_factory=list)
+    resolved_imports: dict[str, str] = Field(default_factory=dict)
 
     # Documentation state management
     documentation_state: str = "pending"  # pending, in_progress, completed, failed
-    dependency_context: dict[str, str] = field(default_factory=dict)
+    dependency_context: dict[str, str] = Field(default_factory=dict)
 
     # Cycle handling
-    cycle_group: Optional[list['ModuleNode']] = None
+    cycle_group: list["ModuleNode"] | None = None
     is_cycle_representative: bool = False
 
-    def add_child(self, child_node: 'ModuleNode') -> None:
+    def add_child(self, child_node: "ModuleNode") -> None:
         """Adds a child module/package to this node."""
         self.children.append(child_node)
 
-    def add_dependency(self, dependency_node: 'ModuleNode') -> None:
+    def add_dependency(self, dependency_node: "ModuleNode") -> None:
         """Adds a dependency relationship to this node."""
         if dependency_node not in self.dependencies:
             self.dependencies.append(dependency_node)
         if self not in dependency_node.dependents:
             dependency_node.dependents.append(self)
 
-    def remove_dependency(self, dependency_node: 'ModuleNode') -> None:
+    def remove_dependency(self, dependency_node: "ModuleNode") -> None:
         """Removes a dependency relationship from this node."""
         if dependency_node in self.dependencies:
             self.dependencies.remove(dependency_node)
         if self in dependency_node.dependents:
             dependency_node.dependents.remove(self)
 
-    def get_all_dependencies(self) -> set['ModuleNode']:
+    def get_all_dependencies(self) -> set["ModuleNode"]:
         """Returns all dependencies (direct and indirect) of this node."""
         all_deps = set()
         for dep in self.dependencies:
@@ -123,7 +135,7 @@ class ModuleNode:
             all_deps.update(dep.get_all_dependencies())
         return all_deps
 
-    def get_all_dependents(self) -> set['ModuleNode']:
+    def get_all_dependents(self) -> set["ModuleNode"]:
         """Returns all dependents (direct and indirect) of this node."""
         all_deps = set()
         for dep in self.dependents:
@@ -149,7 +161,7 @@ class ModuleNode:
         """Returns the full module name (e.g., 'package.submodule')."""
         if self.is_root:
             return self.name
-        # This will be enhanced when we build the full module tree
+        # TODO: This will be enhanced when we build the full module tree
         return self.name
 
     def __hash__(self) -> int:
@@ -165,12 +177,15 @@ class ModuleNode:
     def __repr__(self) -> str:
         """String representation for debugging."""
         deps_count = len(self.dependencies)
-        state = f"({self.documentation_state})" if self.documentation_state != "pending" else ""
+        state = (
+            f"({self.documentation_state})"
+            if self.documentation_state != "pending"
+            else ""
+        )
         return f"ModuleNode(name='{self.name}', is_package={self.is_package}, deps={deps_count}{state})"
 
 
-@dataclass
-class DependencyGraph:
+class DependencyGraph(BaseModel):
     """
     Represents the complete dependency graph of a Python project.
 
@@ -183,10 +198,13 @@ class DependencyGraph:
         cycles: list of detected circular dependency groups
         topological_order: list of nodes in topological order (if computed)
     """
-    nodes: dict[str, ModuleNode] = field(default_factory=dict)
-    root_node: Optional[ModuleNode] = None
-    cycles: list[list[ModuleNode]] = field(default_factory=list)
-    topological_order: list[ModuleNode] = field(default_factory=list)
+
+    model_config = {"arbitrary_types_allowed": True}
+
+    nodes: dict[str, ModuleNode] = Field(default_factory=dict)
+    root_node: ModuleNode | None = None
+    cycles: list[list[ModuleNode]] = Field(default_factory=list)
+    topological_order: list[ModuleNode] = Field(default_factory=list)
 
     def add_node(self, node: ModuleNode) -> None:
         """Adds a node to the graph."""
@@ -194,11 +212,11 @@ class DependencyGraph:
         if node.is_root:
             self.root_node = node
 
-    def get_node(self, path: str) -> Optional[ModuleNode]:
+    def get_node(self, path: str) -> ModuleNode | None:
         """Returns a node by its path."""
         return self.nodes.get(path)
 
-    def get_node_by_name(self, name: str) -> Optional[ModuleNode]:
+    def get_node_by_name(self, name: str) -> ModuleNode | None:
         """Returns a node by its name."""
         for node in self.nodes.values():
             if node.name == name:
@@ -252,8 +270,7 @@ class DependencyGraph:
         return f"DependencyGraph(nodes={node_count}, edges={edge_count}, cycles={cycle_count})"
 
 
-@dataclass
-class DocumentationContext:
+class DocumentationContext(BaseModel):
     """
     Represents the context passed between modules during documentation generation.
 
@@ -266,13 +283,22 @@ class DocumentationContext:
         usage_examples: Dictionary mapping module names to usage examples
         relationships: Dictionary mapping module names to their relationship descriptions
     """
-    module_summaries: dict[str, str] = field(default_factory=dict)
-    interface_descriptions: dict[str, str] = field(default_factory=dict)
-    usage_examples: dict[str, str] = field(default_factory=dict)
-    relationships: dict[str, str] = field(default_factory=dict)
 
-    def add_module_context(self, module_name: str, summary: str, interface: str = "",
-                          examples: str = "", relationships: str = "") -> None:
+    model_config = {"arbitrary_types_allowed": True}
+
+    module_summaries: dict[str, str] = Field(default_factory=dict)
+    interface_descriptions: dict[str, str] = Field(default_factory=dict)
+    usage_examples: dict[str, str] = Field(default_factory=dict)
+    relationships: dict[str, str] = Field(default_factory=dict)
+
+    def add_module_context(
+        self,
+        module_name: str,
+        summary: str,
+        interface: str = "",
+        examples: str = "",
+        relationships: str = "",
+    ) -> None:
         """Adds context information for a specific module."""
         self.module_summaries[module_name] = summary
         if interface:
@@ -285,10 +311,10 @@ class DocumentationContext:
     def get_module_context(self, module_name: str) -> dict[str, str]:
         """Returns all context information for a specific module."""
         return {
-            'summary': self.module_summaries.get(module_name, ''),
-            'interface': self.interface_descriptions.get(module_name, ''),
-            'examples': self.usage_examples.get(module_name, ''),
-            'relationships': self.relationships.get(module_name, '')
+            "summary": self.module_summaries.get(module_name, ""),
+            "interface": self.interface_descriptions.get(module_name, ""),
+            "examples": self.usage_examples.get(module_name, ""),
+            "relationships": self.relationships.get(module_name, ""),
         }
 
     def get_dependency_context_string(self, dependencies: list[ModuleNode]) -> str:
@@ -296,20 +322,30 @@ class DocumentationContext:
         if not dependencies:
             return ""
 
-        context_parts = []
+        # Build dependency context dictionary
+        dep_context_dict: dict[str, str] = {}
         for dep in dependencies:
             dep_name = dep.get_module_name()
             dep_context = self.get_module_context(dep_name)
 
-            if dep_context['summary']:
-                context_parts.append(f"**{dep_name}**: {dep_context['summary']}")
+            # Combine summary and interface if available
+            parts = []
+            if dep_context["summary"]:
+                parts.append(dep_context["summary"])
+            if dep_context["interface"]:
+                parts.append(f"Interface: {dep_context['interface']}")
 
-            if dep_context['interface']:
-                context_parts.append(f"**{dep_name} Interface**: {dep_context['interface']}")
+            if parts:
+                dep_context_dict[dep_name] = " ".join(parts)
 
-        if context_parts:
-            return "\n\n**Dependencies:**\n" + "\n".join(context_parts)
-        return ""
+        if not dep_context_dict:
+            return ""
+
+        # Load and render Jinja template
+        template_dir = Path(__file__).parent / "templates"
+        env = Environment(loader=FileSystemLoader(template_dir))
+        template = env.get_template("dependency_context.j2")
+        return template.render(dependencies=dep_context_dict)
 
     def __repr__(self) -> str:
         """String representation for debugging."""
