@@ -10,6 +10,7 @@ from unittest.mock import Mock, patch
 import pytest
 
 from genai_docs.config import Config
+from genai_docs.exceptions import LLMError
 from genai_docs.llm_client import LLMClient
 
 
@@ -23,26 +24,26 @@ class TestLLMClient:
         self.mock_config.api_key = "test-api-key"
         self.mock_config.model_name = "test-model"
 
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_llm_client_initialization(self, mock_genai):
         """Test LLMClient initialization."""
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
 
         assert client.model is not None
         mock_genai.configure.assert_called_once_with(api_key="test-api-key")
         mock_genai.GenerativeModel.assert_called_once_with("test-model")
 
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_llm_client_initialization_failure(self, mock_genai):
         """Test LLMClient initialization failure."""
         mock_genai.configure.side_effect = Exception("API Error")
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             with pytest.raises(Exception, match="API Error"):
                 LLMClient()
 
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_generate_documentation_success(self, mock_genai):
         """Test successful documentation generation."""
         # Mock the response
@@ -56,14 +57,14 @@ class TestLLMClient:
         mock_model.generate_content.return_value = mock_response
         mock_genai.GenerativeModel.return_value = mock_model
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
             result = client.generate_documentation("Test prompt")
 
         assert result == "Generated documentation"
         mock_model.generate_content.assert_called_once()
 
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_generate_documentation_empty_response(self, mock_genai):
         """Test documentation generation with empty response."""
         # Mock empty response
@@ -74,40 +75,40 @@ class TestLLMClient:
         mock_model.generate_content.return_value = mock_response
         mock_genai.GenerativeModel.return_value = mock_model
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
-            result = client.generate_documentation("Test prompt")
+            with pytest.raises(LLMError) as exc_info:
+                client.generate_documentation("Test prompt")
+            # Should raise LLMError with appropriate message
+            assert "unexpected or empty" in str(exc_info.value)
 
-        assert result.startswith("Error:")
-        assert "unexpected or empty" in result
-
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_generate_documentation_api_error(self, mock_genai):
         """Test documentation generation with API error."""
         mock_model = Mock()
         mock_model.generate_content.side_effect = Exception("API Error")
         mock_genai.GenerativeModel.return_value = mock_model
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
-            result = client.generate_documentation("Test prompt")
+            with pytest.raises(LLMError) as exc_info:
+                client.generate_documentation("Test prompt")
+            # Should raise LLMError with appropriate message
+            assert "API Error" in str(exc_info.value)
 
-        assert result.startswith("Error:")
-        assert "API Error" in result
-
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_generate_documentation_no_model(self, mock_genai):
         """Test documentation generation when model is not configured."""
         mock_genai.GenerativeModel.return_value = None
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
             client.model = None  # Simulate unconfigured model
 
-            with pytest.raises(RuntimeError, match="LLM client not properly configured"):
+            with pytest.raises(LLMError, match="LLM client not properly configured"):
                 client.generate_documentation("Test prompt")
 
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_generate_project_documentation(self, mock_genai):
         """Test project documentation generation."""
         # Mock successful response
@@ -121,7 +122,7 @@ class TestLLMClient:
         mock_model.generate_content.return_value = mock_response
         mock_genai.GenerativeModel.return_value = mock_model
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
 
             project_files = {"pyproject.toml": "[project]\nname = 'test'"}
@@ -136,7 +137,7 @@ class TestLLMClient:
         # Just verify the method was called, skip detailed prompt inspection for now
         assert mock_model.generate_content.called
 
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_generate_package_documentation(self, mock_genai):
         """Test package documentation generation."""
         # Mock successful response
@@ -150,7 +151,7 @@ class TestLLMClient:
         mock_model.generate_content.return_value = mock_response
         mock_genai.GenerativeModel.return_value = mock_model
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
 
             children_docs = ["Child module documentation"]
@@ -167,7 +168,7 @@ class TestLLMClient:
         # Just verify the method was called, skip detailed prompt inspection for now
         assert mock_model.generate_content.called
 
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_generate_module_documentation(self, mock_genai):
         """Test module documentation generation."""
         # Mock successful response
@@ -181,7 +182,7 @@ class TestLLMClient:
         mock_model.generate_content.return_value = mock_response
         mock_genai.GenerativeModel.return_value = mock_model
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
 
             module_content = "def test_function():\n    return True"
@@ -195,7 +196,7 @@ class TestLLMClient:
         # Just verify the method was called, skip detailed prompt inspection for now
         assert mock_model.generate_content.called
 
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_generate_module_documentation_no_content(self, mock_genai):
         """Test module documentation generation without content."""
         # Mock successful response
@@ -209,7 +210,7 @@ class TestLLMClient:
         mock_model.generate_content.return_value = mock_response
         mock_genai.GenerativeModel.return_value = mock_model
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
 
             result = client.generate_module_documentation("test_module", None)
@@ -221,7 +222,7 @@ class TestLLMClient:
         # Just verify the method was called, skip detailed prompt inspection for now
         assert mock_model.generate_content.called
 
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_generate_project_documentation_no_files(self, mock_genai):
         """Test project documentation generation without project files."""
         # Mock successful response
@@ -235,7 +236,7 @@ class TestLLMClient:
         mock_model.generate_content.return_value = mock_response
         mock_genai.GenerativeModel.return_value = mock_model
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
 
             result = client.generate_project_documentation({}, [])
@@ -247,7 +248,7 @@ class TestLLMClient:
         # Just verify the method was called, skip detailed prompt inspection for now
         assert mock_model.generate_content.called
 
-    @patch('genai_docs.llm_client.genai')
+    @patch("genai_docs.llm_client.genai")
     def test_generate_package_documentation_no_children(self, mock_genai):
         """Test package documentation generation without children."""
         # Mock successful response
@@ -261,7 +262,7 @@ class TestLLMClient:
         mock_model.generate_content.return_value = mock_response
         mock_genai.GenerativeModel.return_value = mock_model
 
-        with patch('genai_docs.llm_client.config', self.mock_config):
+        with patch("genai_docs.llm_client.config", self.mock_config):
             client = LLMClient()
 
             result = client.generate_package_documentation("test_package", [], None)
@@ -274,9 +275,9 @@ class TestLLMClient:
         if call_args and call_args.args:
             contents = call_args.args[0]
             assert len(contents) == 1
-            assert contents[0]['role'] == 'user'
-            assert len(contents[0]['parts']) == 1
-            prompt = contents[0]['parts'][0]['text']
+            assert contents[0]["role"] == "user"
+            assert len(contents[0]["parts"]) == 1
+            prompt = contents[0]["parts"][0]["text"]
             assert "test_package" in prompt
             assert "does not contain any direct sub-modules or sub-packages" in prompt
         else:
